requirements-api.txt
# FastAPI stack dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Async task execution
celery==5.3.4
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for testing
httpx==0.25.2
